[refactor] 클래스 분리
- 인사 발령 로직을 별도 서비스로 분리
