+++ momentum-dao-be/src/main/java/com/dao/momentum/organization/employee/command/application/controller/EmployeeCommandController.java
@@ -12,12 +12,12 @@ import com.dao.momentum.organization.employee.command.application.dto.response.A
 import com.dao.momentum.organization.employee.command.application.dto.response.EmployeeCSVResponse;
 import com.dao.momentum.organization.employee.command.application.dto.response.EmployeeInfoUpdateResponse;
 import com.dao.momentum.organization.employee.command.application.dto.response.EmployeeRecordsUpdateResponse;
+import com.dao.momentum.organization.employee.command.application.service.AppointCommandService;
 import com.dao.momentum.organization.employee.command.application.service.CSVService;
 import com.dao.momentum.organization.employee.command.application.service.EmployeeCommandService;
 import com.dao.momentum.organization.employee.exception.EmployeeException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -40,6 +40,7 @@ public class EmployeeCommandController {
     private final EmployeeCommandService employeeService;
     private final FileService fileService;
     private final CSVService csvService;
+    private final AppointCommandService appointCommandService;
 
     @Operation(summary = "사원 등록", description = "관리자는 사원의 정보를 입력하여 사원을 등록할 수 있다.")
     @PostMapping
@@ -117,7 +118,7 @@ public class EmployeeCommandController {
 
         return ResponseEntity.status(HttpStatus.CREATED).body(
                 ApiResponse.success(
-                        employeeService.createAppoint(userDetails, request)
+                        appointCommandService.createAppoint(userDetails, request)
                 )
         );
     }
